Rename go.yml to .github/workflows/go.yml (#14)
* Rename go.yml to .github/workflows/go.yml

why did this ever land in root? I used the gh webUI to make this PR.


* fix broken tests


* pkg: bye felicia


* internal: no root level internal
