+++ ninep/protocol/genout.go
