+++ ninep/ufs/ninep_unix.go
@@ -4,20 +4,29 @@
 // This code is imported from the old ninep repo,
 // with some changes.
 
-// +build windows
+// +build !windows
 
 package ufs
 
 import (
 	"os"
+	"syscall"
 
-	"harvey-os.org/pkg/ninep/protocol"
+	"harvey-os.org/ninep/protocol"
 )
 
 func fileInfoToQID(d os.FileInfo) protocol.QID {
 	var qid protocol.QID
+	sysif := d.Sys()
 
-	qid.Path = uint64(d.ModTime().UnixNano())
+	// on systems with inodes, use it.
+	if sysif != nil {
+		stat := sysif.(*syscall.Stat_t)
+		qid.Path = uint64(stat.Ino)
+	} else {
+		qid.Path = uint64(d.ModTime().UnixNano())
+	}
+
 	qid.Version = uint32(d.ModTime().UnixNano() / 1000000)
 	qid.Type = dirToQIDType(d)
